Add delta mode
This is, in essence, a rate that is less subject to noisy error states
in small rates. If you have something that happens every few seconds but
you still want to alert to make sure it's happening over some period of
time, using --rate is error prone. I believe --delta will be better.

As an example of how this works, you point it at a metric that is a
counter and you tell it how long to look back. The --duration option
specifies 10 minutes by default, then --delta instructs the checker to
calculate the delta between that period and now.

This single value is then checked against the comparator specified and
the thresholds and the result is fired.